+++ syne_tune/optimizer/schedulers/searchers/bayesopt/sklearn/__init__.py
